==> 1. Data_Aggregation.ipynb <==
<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h1 style='margin:10px 5px'> 
Master Thesis Yannik Haller - Data Aggregation
</h1>
</div>

<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h2 style='margin:10px 5px'> 
1. Read in and aggregate/merge all csv-files 
</h2>
</div>

<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h3 style='margin:10px 5px'> 
1.1 Pre-Aggregation 
</h3>
</div>

# Import required packages
import os
import glob
import pandas as pd
import numpy as np

# Change pandas' setting to print out long strings
pd.options.display.max_colwidth = 200

# Set the appropriate working directory
os.chdir('D:\\Dropbox\\MA_data')

# Get a list containing all names of the files we want to load and aggregate
extension = 'csv'
filenames_older      = [i for i in glob.glob('Older/articles*'+'.{}'.format(extension))]
filenames_oct_to_jan = [i for i in glob.glob('Oct-Jan/'+'*_Oct_2020_Jan_2021.{}'.format(extension))]

# Check how many filenames are contained in the created lists
print(len(filenames_older))
print(len(filenames_oct_to_jan))

# Concatenate all files in the list "filenames_older" to one dataframe
agg_older = pd.concat([pd.read_csv(f, usecols = ['source', 'source_full', 'date', 'lang', 'news_section', 'title', 'subtitle', 'article', 'author'],
                                   dtype = {'source': object, 'source_full': object, 'date': object, 'lang': object, 'news_section': object,
                                            'title': object, 'subtitle': object, 'article': object, 'author': object}) for f in filenames_older])

# Rearrange the order of the columns
agg_older = agg_older[['source', 'source_full', 'date', 'lang', 'news_section', 'title', 'subtitle', 'article', 'author']]

# Rename the columns where needed
agg_older.rename(columns = {"source": "so", "source_full": "so_txt", "date": "pubDateTime", "lang": "la", "article": "tx"}, inplace = True)

# Lowercase the entries of the columns 'la' and 'news_section' (to ensure data conformity)
agg_older['la'] = agg_older['la'].str.lower()
agg_older['news_section'] = agg_older['news_section'].str.lower()

# Take a look at the created dataframe
agg_older.head(3)

# Take a look at the shape of the created dataframe
agg_older.shape

# Export the concatenated dataframe 'agg_older' as a csv
agg_older.to_csv("agg_older.csv", index = False, encoding = 'utf-8-sig')

# Remove the dataframe to save RAM
del agg_older

# Concatenate all files in the list "filenames_oct_to_jan" to one dataframe
agg_oct_to_jan = pd.concat([pd.read_csv(f, usecols = ['so', 'so_txt', 'pubDateTime', 'la', 'ru', 'ht', 'ut', 'tx', 'annotation_person'],
                                        dtype = {'so': object, 'so_txt': object, 'pubDateTime': object, 'la': object, 'ru': object,
                                                 'ht': object, 'ut': object, 'tx': object, 'annotation_person': object}) for f in filenames_oct_to_jan])

# Rename the columns where needed
agg_oct_to_jan.rename(columns = {"ru": "news_section", "ht": "title", "ut": "subtitle", "annotation_person": "author"}, inplace = True)

# Lowercase the entries of the columns 'la' and 'news_section' (to ensure data conformity)
agg_oct_to_jan['la'] = agg_oct_to_jan['la'].str.lower()
agg_oct_to_jan['news_section'] = agg_oct_to_jan['news_section'].str.lower()

# Take a look at the created dataframe
agg_oct_to_jan.head(3)

# Take a look at the shape of the created dataframe
agg_oct_to_jan.shape

# Export the concatenated dataframe 'agg_oct_to_jan' as a csv
agg_oct_to_jan.to_csv("agg_oct_to_jan.csv", index = False, encoding = 'utf-8-sig')

# Remove the dataframe to save RAM
del agg_oct_to_jan

<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h3 style='margin:10px 5px'> 
1.2 Overall Aggregation
</h3>
</div>

# Define a list containing the names of the pre-aggregated files we want to merge together
all_filenames = ["agg_older.csv", "agg_oct_to_jan.csv"]

# Concatenate all files in the list to one dataframe
agg_csv = pd.concat([pd.read_csv(f, dtype = {'so': object, 'so_txt': object, 'pubDateTime': object, 'la': object, 'news_section': object,
                                             'title': object, 'subtitle': object, 'tx': object, 'author': object}) for f in all_filenames])
# Give an article-specific unique index
agg_csv.index = np.arange(agg_csv.shape[0])

# Take a look at the aggregated dataframe
agg_csv.head(3)

# Take a look at the shape of the aggregated dataframe
agg_csv.shape

# Check how many missing values are contained in each column
agg_csv.isna().sum()

We observe that the data contains several observations with missing article text. As these observations are useless, we drop them from the data in the following

# Remove observations with missing article text
agg_csv.drop(agg_csv.loc[agg_csv.tx.isnull()].index, inplace = True)

# Inspect the different languages in which the articles are written including article counts per language
np.unique(agg_csv.la, return_counts = True)

We observe that the articles are written in 5 different languages:

- German (de)
- English (en)
- French (fr)
- Italian (it)
- Rhaeto-Romanic (rm)

As the commonly used text processing packages are not available for the Roman language, I decide to remove the according articles from the dataset (which anyway only constitute a negligible fraction of the data).

# Remove articles that are written in Rhaeto-Romanic
agg_csv.drop(agg_csv[agg_csv.la == 'rm'].index, inplace = True)

# Identify the number of duplicates (according to news source, publication date and article text)
n_duplicates = sum(agg_csv.duplicated(subset = ['so','pubDateTime','tx'], keep = 'first'))

# Remove identified duplicates and create a new index (from 0 to n-1)
agg_csv.drop_duplicates(subset = ['so','pubDateTime','tx'], keep = 'first', inplace = True, ignore_index = True)

# Display the number of removed duplicates
print(n_duplicates, "duplicates have been removed.")

# Check how many articles are left per language after removing identified duplicates
np.unique(agg_csv.la, return_counts = True)

# Transform pubDateTime into a date
agg_csv.pubDateTime = pd.to_datetime(agg_csv.pubDateTime, yearfirst = True)

# Take a look at the column types
agg_csv.dtypes

# Check how many missing values are contained in each column
agg_csv.isna().sum()

# Take a look at the shape of the data
agg_csv.shape

<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h2 style='margin:10px 5px'> 
2. Extract the article text from the subtitle where these have been confused by mistake
</h2>
</div>

# Define the index of the according articles
idx_confused_1 = np.arange(59019, 65923+1).tolist()
idx_confused_2 = np.arange(684119, 702950+1).tolist()
idx_confused_2.append(2064755)
idx_confused = np.concatenate((np.array(idx_confused_1), np.array(idx_confused_2))).tolist()

# Check for how many articles the entries of the subtitle and text columns have been confused mistakenly
len(idx_confused)

# Change the entrie of the tx column where needed
agg_csv['tx'].loc[idx_confused] = agg_csv['subtitle'].loc[idx_confused]

<div class="alert alert-info" style="background-color:#5d3a8e; color:white; padding:0px 10px; border-radius:5px;"><h2 style='margin:10px 5px'> 
3. Export the aggregated csv-file 
</h2>
</div>

# Export the concatenated dataframe as a full csv
agg_csv.to_csv("agg_csv.csv", encoding = 'utf-8-sig')

# Export the concatenated dataframe as a sparse csv (containing merely the most crucial columns)
agg_csv.to_csv("agg_csv_sparse.csv", columns = ['so', 'la', 'tx'], encoding = 'utf-8-sig')

# Export the subsample of all articles that are written in the same language separately
agg_csv[agg_csv.la == 'de'].to_csv("agg_csv_sparse_de.csv", columns = ['so', 'la', 'tx'], encoding = 'utf-8-sig')
agg_csv[agg_csv.la == 'en'].to_csv("agg_csv_sparse_en.csv", columns = ['so', 'la', 'tx'], encoding = 'utf-8-sig')
agg_csv[agg_csv.la == 'fr'].to_csv("agg_csv_sparse_fr.csv", columns = ['so', 'la', 'tx'], encoding = 'utf-8-sig')
agg_csv[agg_csv.la == 'it'].to_csv("agg_csv_sparse_it.csv", columns = ['so', 'la', 'tx'], encoding = 'utf-8-sig')

# Read the data back in as follows
#de_tx = pd.read_csv("agg_csv_sparse_de.csv", index_col = 0)
#en_tx = pd.read_csv("agg_csv_sparse_en.csv", index_col = 0)
#fr_tx = pd.read_csv("agg_csv_sparse_fr.csv", index_col = 0)
#it_tx = pd.read_csv("agg_csv_sparse_it.csv", index_col = 0)
